fix: Convert quaternion to euler angles [#93661584]
When switching from euler angles to quaternions, back to euler angles, the
cached _lastEuler values can not be used, but need to be flushed as soon as
switching back to the quaternion representation. Similarily, the quaternion
needs to be converted to euler angles when only setting the rotation along one
axis.