fix: Synchronize Engine time when switching tabs
Previsouly the now() helper function has been used in order to measure the
time not spent on the tab running the Engine.

This resulted into minor precision loss of the high resolution timestamp
passed to updateables.

This commit refactors the engine to exclusively use the rAF timestamp.

As a result, the same timestamp is being preserved between visibility
changes.