feat: Add utilities#assert [#95469256]
DOMRenderer previously had methods for asserting state.

Putting those methods on the prototype makes it impossible for the
minifier to do any kind of dead-code elimination.

This commit adds an assert utility function used in order to check
if a certain condition is met.

Checking for NODE_ENV and not including the passed in error message
in the thrown error should be trivial and only needs to be done once
in the more generic assert function.