refact: Reduce redundancy in WebGLRenderer
Implements a "Registry" class to avoid having to manually keep track of the
stored values in a separate array. As a side effect, this also optimizes the
process of deregistering objects (especially meshes, where Array#splice has
been used before). In the future, the registry could also be used for storing
child nodes.