refact: Remove obsolete data clone
In the current implementation a new typed array is being instantiated on
every subData call using `concat`. This should be factored out by sending
the data as a transferable ArrayBuffer through postMessage, resulting
into a zero-copy transfer. For the time being, we can at least remove the
obsolete cloning of the array using concat.